fixing tests and moving test data
