+++ config/proposer/testdata/bad-prepare-beacon-proposer-config.json
