+++ config/proposer/testdata/default-only-proposer-config.json
