+++ config/proposer/testdata/good-prepare-beacon-proposer-config-multiple.json
