+++ config/proposer/testdata/good-prepare-beacon-proposer-config.yaml
