+++ validator/node/BUILD.bazel
@@ -4,27 +4,19 @@ go_test(
     name = "go_default_test",
     size = "small",
     srcs = ["node_test.go"],
-    data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
         "//cmd:go_default_library",
         "//cmd/validator/flags:go_default_library",
-        "//config/fieldparams:go_default_library",
-        "//config/params:go_default_library",
-        "//config/validator/service:go_default_library",
-        "//consensus-types/validator:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//io/file:go_default_library",
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//validator/accounts:go_default_library",
         "//validator/accounts/wallet:go_default_library",
-        "//validator/db/iface:go_default_library",
         "//validator/db/kv:go_default_library",
-        "//validator/db/testing:go_default_library",
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/remote-web3signer:go_default_library",
-        "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
