+++ validator/node/node_test.go
@@ -3,31 +3,21 @@ package node
 import (
 	"context"
 	"flag"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"path"
 	"path/filepath"
 	"testing"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/prysmaticlabs/prysm/v5/cmd"
 	"github.com/prysmaticlabs/prysm/v5/cmd/validator/flags"
-	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
-	"github.com/prysmaticlabs/prysm/v5/config/params"
-	validatorserviceconfig "github.com/prysmaticlabs/prysm/v5/config/validator/service"
-	"github.com/prysmaticlabs/prysm/v5/consensus-types/validator"
 	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
 	"github.com/prysmaticlabs/prysm/v5/io/file"
 	"github.com/prysmaticlabs/prysm/v5/testing/assert"
 	"github.com/prysmaticlabs/prysm/v5/testing/require"
 	"github.com/prysmaticlabs/prysm/v5/validator/accounts"
 	"github.com/prysmaticlabs/prysm/v5/validator/accounts/wallet"
-	"github.com/prysmaticlabs/prysm/v5/validator/db/iface"
 	"github.com/prysmaticlabs/prysm/v5/validator/db/kv"
-	dbTest "github.com/prysmaticlabs/prysm/v5/validator/db/testing"
 	"github.com/prysmaticlabs/prysm/v5/validator/keymanager"
 	remoteweb3signer "github.com/prysmaticlabs/prysm/v5/validator/keymanager/remote-web3signer"
 	logtest "github.com/sirupsen/logrus/hooks/test"
@@ -343,723 +333,4 @@ func TestWeb3SignerConfig(t *testing.T) {
 			require.DeepEqual(t, tt.want, got)
 		})
 	}
-}
-
-func TestProposerSettings(t *testing.T) {
-	hook := logtest.NewGlobal()
-
-	type proposerSettingsFlag struct {
-		dir        string
-		url        string
-		defaultfee string
-		defaultgas string
-	}
-
-	type args struct {
-		proposerSettingsFlagValues *proposerSettingsFlag
-	}
-	tests := []struct {
-		name                         string
-		args                         args
-		want                         func() *validatorserviceconfig.ProposerSettings
-		urlResponse                  string
-		wantErr                      string
-		wantLog                      string
-		withdb                       func(db iface.ValidatorDB) error
-		validatorRegistrationEnabled bool
-	}{
-		{
-			name: "Happy Path default only proposer settings file with builder settings,",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/default-only-proposer-config.json",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return &validatorserviceconfig.ProposerSettings{
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0xae967917c465db8578ca9024c205720b1a3651A9"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-		},
-		{
-			name: "Happy Path Config file File, bad checksum",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config-badchecksum.json",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0xae967917c465db8578ca9024c205720b1a3651A9"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0xae967917c465db8578ca9024c205720b1a3651A9"),
-						},
-					},
-				}
-			},
-			wantErr: "",
-			wantLog: "is not a checksum Ethereum address",
-		},
-		{
-			name: "Happy Path Config file File multiple fee recipients",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config-multiple.json",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				key2, err := hexutil.Decode("0xb057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7b")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-							},
-						},
-						bytesutil.ToBytes48(key2): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x60155530FCE8a85ec7055A5F8b2bE214B3DaeFd4"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(35000000),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(40000000),
-						},
-					},
-				}
-			},
-			wantErr: "",
-		},
-		{
-			name: "Happy Path Config URL File",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "./testdata/good-prepare-beacon-proposer-config.json",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-			},
-			wantErr: "",
-		},
-		{
-			name: "Happy Path Config YAML file with custom Gas Limit",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config.yaml",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: 40000000,
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  false,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			wantErr: "",
-		},
-		{
-			name: "Happy Path Suggested Fee ",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "0x6e35733c5af9B61374A128e6F85f553aF09ff89A",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: nil,
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-			},
-			wantErr: "",
-		},
-		{
-			name: "Happy Path Suggested Fee , validator registration enabled",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "0x6e35733c5af9B61374A128e6F85f553aF09ff89A",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: nil,
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			wantErr:                      "",
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "Happy Path Suggested Fee , validator registration enabled and default gas",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "0x6e35733c5af9B61374A128e6F85f553aF09ff89A",
-					defaultgas: "50000000",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: nil,
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: 50000000,
-						},
-					},
-				}
-			},
-			wantErr:                      "",
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "Suggested Fee does not Override Config",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config.json",
-					url:        "",
-					defaultfee: "0x6e35733c5af9B61374A128e6F85f553aF09ff89B",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-			},
-			wantErr: "",
-		},
-		{
-			name: "Suggested Fee with validator registration does not Override Config",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config.json",
-					url:        "",
-					defaultfee: "0x6e35733c5af9B61374A128e6F85f553aF09ff89B",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			wantErr:                      "",
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "Enable Builder flag overrides empty config",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config.json",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "Enable Builder flag does override completed builder config",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config.yaml",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(40000000),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "Only Enable Builder flag",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return &validatorserviceconfig.ProposerSettings{
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "No Flags but saved to DB with builder and override removed builder data",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-			},
-			withdb: func(db iface.ValidatorDB) error {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				settings := &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(40000000),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-				return db.SaveProposerSettings(context.Background(), settings)
-			},
-		},
-		{
-			name: "Enable builder flag but saved to DB without builder data now includes builder data",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-							BuilderConfig: &validatorserviceconfig.BuilderConfig{
-								Enabled:  true,
-								GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-						BuilderConfig: &validatorserviceconfig.BuilderConfig{
-							Enabled:  true,
-							GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-						},
-					},
-				}
-			},
-			withdb: func(db iface.ValidatorDB) error {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				settings := &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-				return db.SaveProposerSettings(context.Background(), settings)
-			},
-			validatorRegistrationEnabled: true,
-		},
-		{
-			name: "No flags, but saved to database",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				return &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-			},
-			withdb: func(db iface.ValidatorDB) error {
-				key1, err := hexutil.Decode("0xa057816155ad77931185101128655c0191bd0214c201ca48ed887f6c4c6adf334070efcd75140eada5ac83a92506dd7a")
-				require.NoError(t, err)
-				settings := &validatorserviceconfig.ProposerSettings{
-					ProposeConfig: map[[fieldparams.BLSPubkeyLength]byte]*validatorserviceconfig.ProposerOption{
-						bytesutil.ToBytes48(key1): {
-							FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-								FeeRecipient: common.HexToAddress("0x50155530FCE8a85ec7055A5F8b2bE214B3DaeFd3"),
-							},
-						},
-					},
-					DefaultConfig: &validatorserviceconfig.ProposerOption{
-						FeeRecipientConfig: &validatorserviceconfig.FeeRecipientConfig{
-							FeeRecipient: common.HexToAddress("0x6e35733c5af9B61374A128e6F85f553aF09ff89A"),
-						},
-					},
-				}
-				return db.SaveProposerSettings(context.Background(), settings)
-			},
-		},
-		{
-			name: "No flags set means empty config",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return nil
-			},
-			wantErr: "",
-		},
-		{
-			name: "Bad File Path",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/bad-prepare-beacon-proposer-config.json",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return nil
-			},
-			wantErr: "failed to unmarshal yaml file",
-		},
-		{
-			name: "Both URL and Dir flags used resulting in error",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/good-prepare-beacon-proposer-config.json",
-					url:        "./testdata/good-prepare-beacon-proposer-config.json",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return &validatorserviceconfig.ProposerSettings{}
-			},
-			wantErr: "cannot specify both",
-		},
-		{
-			name: "Bad Gas value in JSON",
-			args: args{
-				proposerSettingsFlagValues: &proposerSettingsFlag{
-					dir:        "./testdata/bad-gas-value-proposer-settings.json",
-					url:        "",
-					defaultfee: "",
-				},
-			},
-			want: func() *validatorserviceconfig.ProposerSettings {
-				return nil
-			},
-			wantErr: "failed to unmarshal yaml file",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			app := cli.App{}
-			set := flag.NewFlagSet("test", 0)
-			if tt.args.proposerSettingsFlagValues.dir != "" {
-				set.String(flags.ProposerSettingsFlag.Name, tt.args.proposerSettingsFlagValues.dir, "")
-				require.NoError(t, set.Set(flags.ProposerSettingsFlag.Name, tt.args.proposerSettingsFlagValues.dir))
-			}
-			if tt.args.proposerSettingsFlagValues.url != "" {
-				content, err := os.ReadFile(tt.args.proposerSettingsFlagValues.url)
-				require.NoError(t, err)
-				srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					w.WriteHeader(200)
-					w.Header().Set("Content-Type", "application/json")
-					_, err := fmt.Fprintf(w, "%s", content)
-					require.NoError(t, err)
-				}))
-				defer srv.Close()
-
-				set.String(flags.ProposerSettingsURLFlag.Name, tt.args.proposerSettingsFlagValues.url, "")
-				require.NoError(t, set.Set(flags.ProposerSettingsURLFlag.Name, srv.URL))
-			}
-			if tt.args.proposerSettingsFlagValues.defaultfee != "" {
-				set.String(flags.SuggestedFeeRecipientFlag.Name, tt.args.proposerSettingsFlagValues.defaultfee, "")
-				require.NoError(t, set.Set(flags.SuggestedFeeRecipientFlag.Name, tt.args.proposerSettingsFlagValues.defaultfee))
-			}
-			if tt.args.proposerSettingsFlagValues.defaultgas != "" {
-				set.String(flags.BuilderGasLimitFlag.Name, tt.args.proposerSettingsFlagValues.defaultgas, "")
-				require.NoError(t, set.Set(flags.BuilderGasLimitFlag.Name, tt.args.proposerSettingsFlagValues.defaultgas))
-			}
-			if tt.validatorRegistrationEnabled {
-				set.Bool(flags.EnableBuilderFlag.Name, true, "")
-			}
-			cliCtx := cli.NewContext(&app, set, nil)
-			validatorDB := dbTest.SetupDB(t, [][fieldparams.BLSPubkeyLength]byte{})
-			if tt.withdb != nil {
-				err := tt.withdb(validatorDB)
-				require.NoError(t, err)
-			}
-			got, err := proposerSettings(cliCtx, validatorDB)
-			if tt.wantErr != "" {
-				require.ErrorContains(t, tt.wantErr, err)
-				return
-			}
-			if tt.wantLog != "" {
-				assert.LogsContain(t, hook,
-					tt.wantLog,
-				)
-			}
-			w := tt.want()
-			require.DeepEqual(t, w, got)
-
-		})
-	}
-}
-
-func Test_ProposerSettingsWithOnlyBuilder_DoesNotSaveInDB(t *testing.T) {
-	app := cli.App{}
-	set := flag.NewFlagSet("test", 0)
-	set.Bool(flags.EnableBuilderFlag.Name, true, "")
-	cliCtx := cli.NewContext(&app, set, nil)
-	validatorDB := dbTest.SetupDB(t, [][fieldparams.BLSPubkeyLength]byte{})
-	got, err := proposerSettings(cliCtx, validatorDB)
-	require.NoError(t, err)
-	_, err = validatorDB.ProposerSettings(cliCtx.Context)
-	require.ErrorContains(t, "no proposer settings found in bucket", err)
-	want := &validatorserviceconfig.ProposerSettings{
-		DefaultConfig: &validatorserviceconfig.ProposerOption{
-			BuilderConfig: &validatorserviceconfig.BuilderConfig{
-				Enabled:  true,
-				GasLimit: validator.Uint64(params.BeaconConfig().DefaultBuilderGasLimit),
-				Relays:   nil,
-			},
-		},
-	}
-	require.DeepEqual(t, want, got)
 }
